Emulates blossom deployment with docker-compose (#130)
* Emulates blossom deployment with docker-compose

* Allow for passing other commands to blossom container image

* Ignore `docker/` misc bin. We move them on build anyway.

Anything in `docker/` is going to be mounted in a different spot, so
things like "I'm unable to find this import you're referencing" mean
nothing.

* Adds hat-tip in README for docker-compose method

The docker approach helps with manual testing, but is not (currently)
built for running automated tests. That would need more effort put into
it, if it's even feasible.