Return user who already claimed a submission
When a submission is claimed and then tried to claim again,
the user who already claimed the submission is returned
along with an error code.

This is the same as df4564be9964d15a41b3b0652a274e200eb463aa, but
this time not pushed directly to master.

This reverts commit 9d606227092a3a3d96c2a53828e8bd5eb79ed46f.
(The revision of the previous commit)